Adding exception handling to the delete post handler to ensure that only the post owner is able to delete the post

+++ MyDocs.Application/Features/Posts/Commands/DeletePost/DeletePostCommandHandler.cs
@@ -1,5 +1,6 @@
 ﻿using AutoMapper;
 using MediatR;
+using MyDocs.Application.Contracts;
 using MyDocs.Application.Contracts.Persistance;
 using MyDocs.Application.Exceptions;
 using MyDocs.Domain.Entities;
@@ -16,19 +17,26 @@ namespace MyDocs.Application.Features.Posts.Commands.DeletePost
     {
         private readonly IMapper _mapper;
         private readonly IAsyncRepository<Post> _postRepository;
+        private readonly ILoggedInUserService _loggedInUserService;
 
-        public DeletePostCommandHandler(IMapper mapper, IAsyncRepository<Post> postRepository)
+        public DeletePostCommandHandler(IMapper mapper, IAsyncRepository<Post> postRepository, ILoggedInUserService loggedInUserService)
         {
             _mapper = mapper;
             _postRepository = postRepository;
+            _loggedInUserService = loggedInUserService;
         }
         public async Task<Unit> Handle(DeletePostCommand request, CancellationToken cancellationToken)
         {
-            var postToDelete = await _postRepository.GetByIdAsync(request.Id);
+            var postToDelete = await _postRepository.GetByIdAsync(request.PostId);
 
             if (postToDelete == null)
             {
-                throw new NotFoundException(nameof(Post), request.Id);
+                throw new NotFoundException(nameof(Post), request.PostId);
+            }
+
+            if (request.UserId != Guid.Parse(_loggedInUserService.UserId))
+            {
+                throw new UnauthorizedActionException("You must own this post to delete it");
             }
 
             await _postRepository.DeleteAsync(postToDelete);
